Moved new fixture to conftest within test/indexes
